+++ catalog-rest-service/src/main/java/org/openmetadata/catalog/resources/pipelines/PipelineResource.java
@@ -416,6 +416,7 @@ public class PipelineResource {
       })
   public Response delete(@Context UriInfo uriInfo, @Context SecurityContext securityContext, @PathParam("id") String id)
       throws IOException {
+    SecurityUtil.checkAdminOrBotRole(authorizer, securityContext);
     DeleteResponse<Pipeline> response = dao.delete(securityContext.getUserPrincipal().getName(), id);
     return response.toResponse();
   }
